Rename to `Explorer.tsx`
See also: #8
